feat(data mgr): Link abstraction acts as an Observable source for consumers.
A Link is a bit like a StreamController in Dart - it acts as a sink that feeds
into an Observable source which will be returned to consumers. A Link provides
a hot Observable and tracks its subscribers. It is capable of notifying when
all subscribers have disconnected, allowing for cleanup when an application is
no longer interested in the data it has been providing.