fix(idb): Use an Observable instead of a ReplaySubject.
Any consumer of Tactical will have to supply a copy of the 'rx' library, which has several different distributions. These distributions bundle some subset of RxJS functionality (so one does not need to import the entire library to use only the basics). Unfortunately, ReplaySubject is not included in the most common distribution bundle.

Angular distributes this common bundle of RxJS by default, so consumers using Angular do not have ReplaySubject available. This change makes it possible to use Angular's distribution of RxJS with Tactical by removing the dependency on ReplaySubject.